+++ common/src/main/java/com/kryeit/compat/CompatAddon.java
@@ -4,7 +4,6 @@ import com.kryeit.utils.Utils;
 
 public enum CompatAddon {
     CREATE_DECO("createdeco"),
-    BREWIN_AND_CHEWIN("brewinandchewin"),
     LUCKPERMS("luckperms");
 
     private final String id;