+++ changelog.md
@@ -1,5 +1,7 @@
 # 0.3.2
 - Now train, dive, vote and statistic-based missions can only appear once
+- Deleted drinks from Eating mission, and created a new Drinking mission.
+- Added a new tag, `#c:drinks`
 
 # 0.2.7
 - Added compat with Create: Missions' addons