+++ changelog.md
@@ -1,7 +1,7 @@
 # 0.3.2
 - Now train, dive, vote and statistic-based missions can only appear once
 - Deleted drinks from Eating mission, and created a new Drinking mission.
-- Added a new tag, `#c:drinks`
+- Added new tags, `#c:drinks` and `#brewinandchewin:drinks`
 - Added support for tags in missions.json
 
 # 0.2.7